+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Unreleased
 
+- added more decoder integration tests (com -> elpm)
+
 - moved sfr_lookup[] table to a separate file. 
 - adjusting readme ;)